+++ src/main/java/expresscarts/mixin/NewMinecartBehaviorMixin.java
@@ -1,10 +1,13 @@
 package expresscarts.mixin;
 
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import expresscarts.ExpressCartsConfig;
 import expresscarts.ExpressMinecartEntity;
 import net.minecraft.world.entity.vehicle.AbstractMinecart;
 import net.minecraft.world.entity.vehicle.MinecartBehavior;
 import net.minecraft.world.entity.vehicle.NewMinecartBehavior;
+import net.minecraft.world.phys.Vec3;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -39,4 +42,15 @@ public abstract class NewMinecartBehaviorMixin extends MinecartBehavior {
             cir.setReturnValue((baseSpeed * blockMultiplier * waterSpeedMultiplier) / 20.0);
         }
     }
+
+    @WrapOperation(method = "calculateHaltTrackSpeed(Lnet/minecraft/world/phys/Vec3;Lnet/minecraft/world/level/block/state/BlockState;)Lnet/minecraft/world/phys/Vec3;", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/phys/Vec3;scale(D)Lnet/minecraft/world/phys/Vec3;"))
+    // Modify the scale applied to the cart's velocity when passing over an unpowered powered rail
+    private Vec3 unpoweredPoweredRailVelocityScale(Vec3 instance, double scale, Operation<Vec3> original) {
+        if (this.minecart instanceof ExpressMinecartEntity && ExpressCartsConfig.fastUnpoweredSlowdown) {
+            return original.call(instance, 0.4); // seems reasonable in testing. stops similar or faster than vanilla carts (0.5) even at high max speeds (>128) with a long runway
+        } else {
+            return original.call(instance, scale);
+        }
+    }
+
 }